Separate out 'adapt to html' code from annotation transformer
