+++ rosa-iiif/rosa-iiif-presentation-core/src/main/java/rosa/iiif/presentation/core/util/MarginaliaHtmlAdapter.java
@@ -1,0 +1,109 @@
+package rosa.iiif.presentation.core.util;
+
+import com.google.inject.Inject;
+import org.apache.commons.lang3.StringEscapeUtils;
+import rosa.archive.model.Book;
+import rosa.archive.model.BookCollection;
+import rosa.archive.model.BookImage;
+import rosa.archive.model.aor.InternalReference;
+import rosa.archive.model.aor.Location;
+import rosa.archive.model.aor.Marginalia;
+import rosa.archive.model.aor.MarginaliaLanguage;
+import rosa.archive.model.aor.Position;
+import rosa.archive.model.aor.XRef;
+import rosa.iiif.presentation.core.PresentationUris;
+import rosa.iiif.presentation.core.jhsearch.JHSearchField;
+
+import javax.xml.stream.XMLStreamException;
+import java.util.ArrayList;
+import java.util.List;
+
+public class MarginaliaHtmlAdapter extends AnnotationBaseHtmlAdapter<Marginalia> {
+
+    @Inject
+    public MarginaliaHtmlAdapter(PresentationUris pres_uris) {
+        super(pres_uris);
+    }
+
+    @Override
+    Class<Marginalia> getAnnotationType() {
+        return Marginalia.class;
+    }
+
+    @Override
+    void annotationAsHtml(BookCollection col, Book book, BookImage page, Marginalia annotation) throws XMLStreamException{
+        List<String> transcription = new ArrayList<>();
+        List<String> people = new ArrayList<>();
+        List<String> books = new ArrayList<>();
+        List<String> locs = new ArrayList<>();
+        List<String> symb = new ArrayList<>();
+
+        // Left, top, right, bottom
+        boolean[] orientation = new boolean[4];
+        List<Location> positions = new ArrayList<>();
+        List<XRef> xrefs = new ArrayList<>();
+        List<InternalReference> iRefs = new ArrayList<>();
+
+        for (MarginaliaLanguage lang : annotation.getLanguages()) {
+            for (Position pos : lang.getPositions()) {
+                transcription.addAll(pos.getTexts());
+                people.addAll(pos.getPeople());
+                books.addAll(pos.getBooks());
+                locs.addAll(pos.getLocations());
+                symb.addAll(pos.getSymbols());
+
+                xrefs.addAll(pos.getxRefs());
+                iRefs.addAll(pos.getInternalRefs());
+
+                // No default case. If orientation is not 0, 90, 180, 270 then do nothing
+                switch (pos.getOrientation()) {
+                    case 0:
+                        orientation[1] = true;
+                        break;
+                    case 90:
+                        orientation[0] = true;
+                        break;
+                    case 180:
+                        orientation[3] = true;
+                        break;
+                    case 270:
+                        orientation[2] = true;
+                        break;
+                }
+
+                // Add icon for position(s) on page
+                positions.add(pos.getPlace());
+            }
+        }
+
+        writer.writeStartElement("p");
+
+        // ------ Add orientation + location icons ------
+        assembleLocationIcon(orientation, positions.toArray(new Location[0]), writer);
+
+        if (isNotEmpty(annotation.getOtherReader())) {
+            writer.writeStartElement("p");
+            writer.writeCharacters("Reader: " + annotation.getOtherReader());
+            writer.writeEndElement();
+        }
+
+        // Add transcription
+        writer.writeStartElement("p");
+        writer.writeCharacters(StringEscapeUtils.escapeHtml4(transcription.toString()));
+        writer.writeEndElement();
+
+        // Add translation
+        addTranslation(annotation.getTranslation(), writer);
+
+        addListOfValues("Symbols:", symb, writer);
+        addSearchableList("People:", people, JHSearchField.PEOPLE, pres_uris.getCollectionURI(col.getId()), writer);
+        addListOfValues("Books:", books, writer);
+        addListOfValues("Locations:", locs, writer);
+
+        // Add list of X-refs
+        addXRefs(xrefs, writer);
+        addInternalRefs(col, iRefs, writer);
+
+        writer.writeEndElement();
+    }
+}
