+++ rosa-iiif/rosa-iiif-presentation-core/src/test/java/rosa/iiif/presentation/core/ArchiveIIIFPresentationServiceTest.java
@@ -6,27 +6,16 @@ import static org.junit.Assert.assertTrue;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Set;
 
 import org.json.JSONObject;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import rosa.archive.core.ArchiveNameParser;
 import rosa.archive.core.BaseSearchTest;
 import rosa.iiif.presentation.core.transform.PresentationTransformer;
-import rosa.iiif.presentation.core.transform.Transformer;
-import rosa.iiif.presentation.core.transform.impl.AnnotationListTransformer;
-import rosa.iiif.presentation.core.transform.impl.AnnotationTransformer;
-import rosa.iiif.presentation.core.transform.impl.CanvasTransformer;
 import rosa.iiif.presentation.core.transform.impl.CollectionTransformer;
 import rosa.iiif.presentation.core.transform.impl.JsonldSerializer;
-import rosa.iiif.presentation.core.transform.impl.ManifestTransformer;
 import rosa.iiif.presentation.core.transform.impl.PresentationTransformerImpl;
-import rosa.iiif.presentation.core.transform.impl.RangeTransformer;
-import rosa.iiif.presentation.core.transform.impl.SequenceTransformer;
-import rosa.iiif.presentation.core.transform.impl.TransformerSet;
 import rosa.iiif.presentation.model.PresentationRequest;
 import rosa.iiif.presentation.model.PresentationRequestType;
 
@@ -54,20 +43,9 @@ public class ArchiveIIIFPresentationServiceTest extends BaseSearchTest {
         ImageIdMapper imageIdMapper = new JhuImageIdMapper(new HashMap<String, String>());
 
         CollectionTransformer collectionTransformer = new CollectionTransformer(requestFormatter, simpleStore, imageFormatter, imageIdMapper);
-        CanvasTransformer canvasTransformer = new CanvasTransformer(requestFormatter, imageFormatter, imageIdMapper);
-        SequenceTransformer sequenceTransformer = new SequenceTransformer(requestFormatter, canvasTransformer);
-        AnnotationTransformer annotationTransformer = new AnnotationTransformer(requestFormatter, new ArchiveNameParser());
 
-        Set<Transformer<?>> transformers = new HashSet<>();
-        transformers.add(new AnnotationListTransformer(requestFormatter, annotationTransformer));
-        transformers.add(canvasTransformer);
-        transformers.add(sequenceTransformer);
-        transformers.add(new ManifestTransformer(requestFormatter, sequenceTransformer, new RangeTransformer(requestFormatter)));
-        transformers.add(new RangeTransformer(requestFormatter));
-
-        TransformerSet transformerSet = new TransformerSet(transformers);
-
-        PresentationTransformer transformer = new PresentationTransformerImpl(requestFormatter, transformerSet,
+        PresentationTransformer transformer = new PresentationTransformerImpl(requestFormatter,
+                PresentationTestUtils.transformerSet(requestFormatter, imageFormatter, imageIdMapper),
                 collectionTransformer);
 
         service = new ArchiveIIIFPresentationService(simpleStore, serializer, transformer, 1000);
